refactor: types of auth consolidated into a centralized file
